feat: :sparkles: Creation of the logical component and the view for the route /customer/products
In this commit, we also made changes to the global store using Zustand + Immer, in this way, functions that can manipulate the base state, are within the store. The main reason of this, is to make them as generic as possible for handling different data changes coming from different tags.